[REFACTOR] move all the components to corresponding subdirectories
[why]
Every component was created under Components directory, so I moved them to subdirectories.

[what]
Components/

[etc]
