+++ Frontend/Views/Components/IssueDetail/CommentWrapper.js
