+++ Frontend/Views/Components/IssueDetail/IssueDetailTitle.js
