+++ Frontend/Views/Pages/IssueDetailPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import IssueDetailWrapper from '../Components/IssueDetailWrapper';
+import IssueDetailWrapper from '../Components/IssueDetail/IssueDetailWrapper';
 
 const IssueDetailPage = () => {
   const { issueId } = useParams();
